// src/app/utils/models/evidencias-archivos-add.ts
export class EvidenciasArchivosAdd {
    evidenciasId?: any;
    indicadorSiacId: number;
    archivoId: number;
    evidenciaId: number;
    nombre: string;
    activo: boolean;
    usuarioCreacion: number;
    File: string;
    tipoEvidencia: string;
    justificacion: string;
    url: string;
    nivelModalidadId: number;
}

//   interface EvidenciasArchivosAdd {
//     evidenciasId?: any;
//     indicadorSiacId: number;
//     archivoId: number;
//     evidenciaId: number;
//     nombre: string;
//     activo: boolean;
//     usuarioCreacion: number;
//     File: string;
//   }